Fix a tag event listener on external pages
The scrolling behavior being implemented in this function causes anchor tags to outside pages to break. For example, in my case I was linking to a section on another page on my site like so: <a href="/contact/#contact">Link to contact page, contact form section</a>

Implementing this if statement will add a check that will allow this behavior to work, whereas it previously shot a console error. It still retains the original scrolling animation functionality for anchor tags inside the same page.

Test this out and implement it, should be pretty straightforward. Had me stumped for a few minutes.